Video recorder (#11)
* Initial video recording

* Video generation improvements

* Update README.md

* Unify audio and video events as FrameCompleted

* Refactor audio/video recording code

* Scaling and cropping

* File clean up, audio options

* Update readme

* Recording preferences

* Restructure recorder project

* Rename to MediaRecorder

* Remove unused using

* Add notification

* FFmpeg path and check

+++ src/Spectron.Emulation/Snapshot/SzxFileExtensions.cs
@@ -48,7 +48,7 @@ public static class SzxFileExtensions
                             attributeData.Paper;
                     }
 
-                    Buffer[bufferIndex + bit] = color.Abgr;
+                    Buffer[bufferIndex + bit] = (int)color.Abgr;
                 }
             }
         }
